integration tests using TestContainers fixes #11
