+++ .travis.yml
@@ -1,0 +1,7 @@
+language: java
+jdk:
+  - openjdk8
+
+sudo: required
+services:
+  - docker
